+++ services/authentication/src/Authentication.API/Models/SigninRequest.cs
@@ -1,0 +1,5 @@
+namespace Authentication.API.Models;
+
+public record SigninRequest(
+    string Email,
+    string Password);